+++ pysui/sui/sui_constants.py
@@ -14,27 +14,28 @@
 
 """Sui Constants."""
 
+SUI_COIN_DENOMINATOR: int = 1000000000
 
-SUI_HEX_ADDRESS_STRING_LEN = 42
+SUI_HEX_ADDRESS_STRING_LEN: int = 42
 """Length of valid address string with '0x' prefix."""
-SUI_ADDRESS_STRING_LEN = 40
+SUI_ADDRESS_STRING_LEN: int = 40
 """Length of valid address string without '0x' prefix."""
 
-SUI_KEYPAIR_LEN = 88
+SUI_KEYPAIR_LEN: int = 88
 """Length of keypair string."""
 
 # ED25519 from keystring bytes
-ED25519_KEYPAIR_BYTES_LEN = 64
+ED25519_KEYPAIR_BYTES_LEN: int = 64
 """Length of ED25519 keypair bytes."""
-ED25519_PUBLICKEY_BYTES_LEN = 32
+ED25519_PUBLICKEY_BYTES_LEN: int = 32
 """Length of ED25519 public key bytes."""
-ED25519_PRIVATEKEY_BYTES_LEN = 32
+ED25519_PRIVATEKEY_BYTES_LEN: int = 32
 """Length of ED25519 private key bytes."""
 
 # SECP256K1 from address bytes
-SECP256K1_KEYPAIR_BYTES_LEN = 65
+SECP256K1_KEYPAIR_BYTES_LEN: int = 65
 """Length of SECP256K1 keypair bytes."""
-SECP256K1_PUBLICKEY_BYTES_LEN = 33
+SECP256K1_PUBLICKEY_BYTES_LEN: int = 33
 """Length of SECP256K1 public key bytes."""
-SECP256K1_PRIVATEKEY_BYTES_LEN = 32
+SECP256K1_PRIVATEKEY_BYTES_LEN: int = 32
 """Length of SECP256K1 private key bytes."""